+++ opencti-platform/opencti-front/lang/front/de.json
@@ -1468,9 +1468,9 @@
   "Filters are not fully displayed": "Filter werden nicht vollständig angezeigt",
   "Find examples on our documentation.": "Beispiele finden Sie auf unserem {link}.",
   "Fintel design": "Fintel-Muster",
-  "Fintel Design": "Fintel Design",
   "Fintel design | Customization | Settings": "Fintel Design | Anpassung | Einstellungen",
   "Fintel Designs": "Fintel Designs",
+  "Fintel designs": "Fintel-Entwürfe",
   "FINTEL template created": "FINTEL-Vorlage erstellt",
   "FINTEL Templates": "FINTEL-Vorlagen",
   "FINTEL templates are available with an Enterprise Edition subscription": "FINTEL Vorlagen sind mit einem Enterprise Edition Abonnement verfügbar",
