+++ opencti-platform/opencti-front/lang/front/en.json
@@ -1468,9 +1468,9 @@
   "Filters are not fully displayed": "Filters are not fully displayed",
   "Find examples on our documentation.": "Find examples on our {link}.",
   "Fintel design": "Fintel design",
-  "Fintel Design": "Fintel Design",
   "Fintel design | Customization | Settings": "Fintel design | Customization | Settings",
   "Fintel Designs": "Fintel Designs",
+  "Fintel designs": "Fintel designs",
   "FINTEL template created": "FINTEL template created",
   "FINTEL Templates": "FINTEL Templates",
   "FINTEL templates are available with an Enterprise Edition subscription": "FINTEL templates are available with an Enterprise Edition subscription",
