+++ opencti-platform/opencti-front/lang/front/es.json
@@ -1468,9 +1468,9 @@
   "Filters are not fully displayed": "Los filtros no se muestran completamente",
   "Find examples on our documentation.": "Encuentra ejemplos en nuestro {link}.",
   "Fintel design": "Diseño fintel",
-  "Fintel Design": "Diseño Fintel",
   "Fintel design | Customization | Settings": "Fintel design | Personalización | Configuración",
   "Fintel Designs": "Diseños Fintel",
+  "Fintel designs": "Diseños Fintel",
   "FINTEL template created": "Plantilla FINTEL creada",
   "FINTEL Templates": "Plantillas FINTEL",
   "FINTEL templates are available with an Enterprise Edition subscription": "Las plantillas FINTEL están disponibles con una suscripción Enterprise Edition",
