+++ opencti-platform/opencti-front/lang/front/fr.json
@@ -1468,9 +1468,9 @@
   "Filters are not fully displayed": "Les filtres ne sont pas entièrement affichés",
   "Find examples on our documentation.": "Trouvez des exemples sur notre {link}.",
   "Fintel design": "Fintel design",
-  "Fintel Design": "Fintel Design",
   "Fintel design | Customization | Settings": "Fintel design | Personnalisation | Paramètres",
   "Fintel Designs": "Fintel Designs",
+  "Fintel designs": "Modèles Fintel",
   "FINTEL template created": "Création d'un modèle FINTEL",
   "FINTEL Templates": "Modèles FINTEL",
   "FINTEL templates are available with an Enterprise Edition subscription": "Les modèles FINTEL sont disponibles avec un abonnement à l'édition Entreprise",
@@ -3481,7 +3481,7 @@
   "Update a dissemination list": "Mise à jour d'une liste de diffusion",
   "Update a feed": "Modifier un flux",
   "Update a feedback": "Mettre à jour un commentaire",
-  "Update a Fintel design": "Mise à jour d'un dessin ou modèle Fintel",
+  "Update a Fintel design": "Mise à jour d'un modèle Fintel",
   "Update a group": "Modifier un groupe",
   "Update a grouping": "Modifier un groupement",
   "Update a individual": "Modifier un individu",
