+++ opencti-platform/opencti-front/lang/front/it.json
@@ -1469,6 +1469,8 @@
   "Find examples on our documentation.": "Trova esempi nella nostra {link}.",
   "Fintel design": "Design Fintel",
   "Fintel design | Customization | Settings": "Fintel design | Personalizzazione | Impostazioni",
+  "Fintel Designs": "Progetti Fintel",
+  "Fintel designs": "Disegni Fintel",
   "FINTEL template created": "Template FINTEL creato",
   "FINTEL Templates": "Template FINTEL",
   "FINTEL templates are available with an Enterprise Edition subscription": "I template FINTEL sono disponibili con un abbonamento all'edizione Enterprise",
@@ -3479,6 +3481,7 @@
   "Update a dissemination list": "Aggiorna una lista di diffusione",
   "Update a feed": "Aggiorna un feed",
   "Update a feedback": "Aggiorna un feedback",
+  "Update a Fintel design": "Aggiornare un progetto Fintel",
   "Update a group": "Aggiorna un gruppo",
   "Update a grouping": "Aggiorna un raggruppamento",
   "Update a individual": "Aggiorna un individuo",
