+++ opencti-platform/opencti-front/lang/front/ja.json
@@ -1468,9 +1468,9 @@
   "Filters are not fully displayed": "フィルターが完全に表示されていません",
   "Find examples on our documentation.": "私たちの{link}で例を見てください。",
   "Fintel design": "フィンテルデザイン",
-  "Fintel Design": "フィンテルデザイン",
   "Fintel design | Customization | Settings": "フィンテルデザイン｜カスタマイズ｜設定",
   "Fintel Designs": "フィンテルデザイン",
+  "Fintel designs": "フィンテルのデザイン",
   "FINTEL template created": "FINTELテンプレート作成",
   "FINTEL Templates": "FINTELテンプレート",
   "FINTEL templates are available with an Enterprise Edition subscription": "FINTELテンプレートはEnterprise Editionのサブスクリプションでご利用いただけます。",
