+++ opencti-platform/opencti-front/lang/front/ko.json
@@ -1468,9 +1468,9 @@
   "Filters are not fully displayed": "필터가 완전히 표시되지 않습니다",
   "Find examples on our documentation.": "{link}에서 예시를 찾아보세요.",
   "Fintel design": "핀텔 디자인",
-  "Fintel Design": "핀텔 디자인",
   "Fintel design | Customization | Settings": "핀텔 디자인 | 사용자 지정 | 설정",
   "Fintel Designs": "핀텔 디자인",
+  "Fintel designs": "핀텔 디자인",
   "FINTEL template created": "핀텔 템플릿 생성",
   "FINTEL Templates": "FINTEL 템플릿",
   "FINTEL templates are available with an Enterprise Edition subscription": "FINTEL 템플릿은 엔터프라이즈 에디션 구독을 통해 사용할 수 있습니다",
