+++ opencti-platform/opencti-front/lang/front/zh.json
@@ -1468,9 +1468,9 @@
   "Filters are not fully displayed": "筛选器未完全显示",
   "Find examples on our documentation.": "在我们的 {link} 中查找示例。",
   "Fintel design": "芬特尔设计",
-  "Fintel Design": "芬特尔设计公司",
   "Fintel design | Customization | Settings": "Fintel 设计 | 个性化 | 设置",
   "Fintel Designs": "芬特尔设计公司",
+  "Fintel designs": "芬特尔设计",
   "FINTEL template created": "创建 FINTEL 模板",
   "FINTEL Templates": "FINTEL 模板",
   "FINTEL templates are available with an Enterprise Edition subscription": "订购企业版可获得 FINTEL 模板",
